+++ .changeset/hungry-bottles-flow.md
@@ -1,0 +1,7 @@
+---
+'@clerk/nextjs': patch
+'@clerk/clerk-react': patch
+'@clerk/remix': patch
+---
+
+Export `EmailLinkErrorCode` from `/errors` module
