+++ .changeset/light-buckets-check.md
@@ -1,0 +1,5 @@
+---
+'@clerk/remix': patch
+---
+
+Introduce `/errors` module. This path exports all error-related APIs such as `isClerkAPIResponseError`, `isEmailLinkError`, `isKnownError`, `isMetamaskError`, `EmailLinkErrorCode`
