+++ packages/remix/api.server/package.json
@@ -1,0 +1,4 @@
+{
+  "main": "../dist/api/index.js",
+  "types": "../dist/api/index.d.ts"
+}
