+++ packages/remix/package.json
@@ -36,6 +36,10 @@
     "./api.server": {
       "types": "./dist/api/index.d.ts",
       "default": "./dist/api/index.js"
+    },
+    "./errors": {
+      "types": "./dist/errors.d.ts",
+      "default": "./dist/errors.js"
     }
   },
   "main": "dist/index.js",
