+++ packages/remix/ssr.server/package.json
@@ -1,0 +1,4 @@
+{
+  "main": "../dist/ssr/index.js",
+  "types": "../dist/ssr/index.d.ts"
+}
